// font_end_khachhang/src/types/product.type.ts
export interface CustomError {
  message: string;
  response?: {
    data?: string;
    status?: number;
  };
}
// ...
export interface IBill {
  id: number;
  type: number;
  address: string;
  index: number;
  status: number;
  phoneNumber: string;
  voucher: any;
  customer: any;
  moneyShip: number;
  code: string;
  note: string;
  moneyReduce: number;
  totalMoney: number;
  employee: any;
  payDate: any;
  shipDate: any;
  desiredDate: any;
  receiveDate: any;
  createAt: string;
  customerName: string;
}
// ...
export interface IDetailOrder {
  name: string;
  id: number;
  size: string;
  index: number;
  images: string;
  quantity: number;
  price: number;
  sole: string;
  color: string;
  discountValue: number;
  shoeCode: string;
  discountPercent: number;
  shoePrice: number;
}